+++ CHANGELOG.md
@@ -3,8 +3,14 @@ All notable changes to this project are documented in this file.
 
 ## 0.2 - 2015-02-04
 ### Changed
+- Add `assets.writer` and `assets.command` services to Silex provider
+- Allow Silex provider to accept multiple paths for assets
+- Added `write_on_compile` option for Silex provider
+- Finished README
+- Added `.gitattributes`
 - Cleaned up code comments
 - Added additional unit tests
+
 
 ## 0.1 - 2014-12-30
 ### Added