+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project are documented in this file.
 
 ## 0.2 - 2015-02-04
 ### Changed
+- Added `getMimeTypes()` method to `AssetController` base class so that MIMEs can be modified at runtime
 - Add `assets.writer` and `assets.command` services to Silex provider
 - Allow Silex provider to accept multiple paths for assets
 - Added `write_on_compile` option for Silex provider